+++ integrations/rocket/src/lib.rs
@@ -214,9 +214,9 @@ impl<'r> Responder<'r, 'static> for Response {
             if let Some(cache_control) = self.0.cache_control().value() {
                 response.set_header(Header::new("cache-control", cache_control));
             }
-            for (name, value) in self.0.http_headers() {
-                response.adjoin_header(Header::new(name.to_string(), value.to_string()));
-            }
+        }
+        for (name, value) in self.0.http_headers() {
+            response.adjoin_header(Header::new(name.to_string(), value.to_string()));
         }
 
         response.set_sized_body(body.len(), Cursor::new(body));